Make config.py more declarative
